requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
markdown2==2.4.0
MarkupSafe==1.1.1
Pygments==2.9.0
PyYAML==5.4.1
Werkzeug==1.0.1
